Register es6-promise to any-promise
The package any-promise is a dependency for probe-image-size. Unfortunately
the package is not prepared to correctly pick up the native Promise
implementation from the global object instead it tries to load it from the
window, which obviously doesn't work.

In order to make it work (for now) the es6-promise implementation is
registered as a Promise implementation.