Show images defined in constant classes
By doing a definition lookup via the command
'vscode.executeDefinitionProvider' for statements which look like a
property access, the returned location object could be used to check for
an image definition string.

As this could change across languages (syntax of property access) it
couldn't be generalized without having an API in vscode which grants
access to e.g. the tokenization result or the AST.

In the current implementation the property access is discovered with a
regular expression which accepts the dot (. for C like languages) and the
dollar sign ($ for e.g. SCSS) and a typical variable name.

This resolves #60